style: update prettier and pretty-quick to the current version
That's the latest for pretty-quick, but one version back for prettier
(2.8.8 rather than 3.0.0) because pretty-quick isn't compatible with
prettier 3.0.0 yet.

The change is motivated by the fact that our previous version, prettier
1.19.1, does not support the named tuple syntax we're starting to use
for ResponseMorpheme.

Also, configure prettier to not add trailing commas, because that's
going to change a lot of lines of code.